chore: rename dir from `assistant` to plural
