+++ api/server/routes/assistants/assistants.js
@@ -2,7 +2,7 @@ const multer = require('multer');
 const express = require('express');
 const { FileContext, EModelEndpoint } = require('librechat-data-provider');
 const { updateAssistant, getAssistants } = require('~/models/Assistant');
-const { initializeClient } = require('~/server/services/Endpoints/assistant');
+const { initializeClient } = require('~/server/services/Endpoints/assistants');
 const { getStrategyFunctions } = require('~/server/services/Files/strategies');
 const { uploadImageBuffer } = require('~/server/services/Files/process');
 const { deleteFileByFilter } = require('~/models/File');
