+++ api/server/services/Endpoints/assistants/initializeClient.js
