+++ api/server/services/Endpoints/assistants/initializeClient.spec.js
